- optimized for 32 bit compile

+++ src/files.h
@@ -25,12 +25,12 @@ void files_shutdown();
 struct x16file *x16open(const char *path, const char *attribs);
 void            x16close(struct x16file *f);
 
-int64_t x16size(struct x16file *f);
-int     x16seek(struct x16file *f, int64_t pos, int origin);
-int64_t x16tell(struct x16file *f);
+size_t x16size(struct x16file *f);
+int    x16seek(struct x16file *f, size_t pos, int origin);
+size_t x16tell(struct x16file *f);
 
 int     x16write8(struct x16file *f, uint8_t val);
 uint8_t x16read8(struct x16file *f);
 
-uint64_t x16write(struct x16file *f, const void *data, uint64_t data_size, uint64_t data_count);
-uint64_t x16read(struct x16file *f, void *data, uint64_t data_size, uint64_t data_count);
+size_t x16write(struct x16file *f, const void *data, size_t data_size, size_t data_count);
+size_t x16read(struct x16file *f, void *data, size_t data_size, size_t data_count);
